+++ src/main/java/ru/itpark/Main.java
@@ -3,8 +3,8 @@ package ru.itpark;
 public class Main {
     public static void main(String[] args) {
 
-        double purchaseAmount = 1_900;
-        double lastTotalAmount = 65_000;
+        int purchaseAmount = 1_900;
+        int lastTotalAmount = 65_000;
         BonusService bonusService = new BonusService();
         int bonus = bonusService.calculateBonus(purchaseAmount, lastTotalAmount);
